Markdown: fix overlapping highlight of alert words and mentions.
Added regex to the process_message function so that anything
in a span tag (which represents a mention or an alert word
that has already been marked) is ignored.

This prevents creation of nested span tags so only one
alert word can be applied to a section of the message.
And also prevents usernames in mentions from being
highlighted, since they are also contained in span tags.

Tests were added to cover the behavior described in
the issues page.

Fixes #9157.